# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MagickVoice Sentinel API
  version: 1.0.0
  description: |
    Source of truth for the Sentinel REST surface. Client, gateway and web all
    generate types from this file; a change to an API shape starts here.

    **Authorisation model.** Every route except `/v1/devices/enroll` and `/healthz`
    requires a Google Cloud Identity Platform ID token as a bearer credential.
    `tenant_id`, `role` and `team_id` are read from the *verified token claims* and
    never from a request body or path parameter. Endpoints under `/v1/me` derive the
    subject UID from the token; a UID supplied by the caller is ignored.

    Device-scoped routes (`/v1/policy`, `/v1/heartbeat`, `/v1/ingest`) additionally
    require mutual TLS with the certificate issued at enrollment.
servers:
  - url: https://api.sentinel.magickvoice.com
    description: production (ap-south-1)
  - url: http://localhost:8080
    description: local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: Calls visible at the caller's scope
      description: |
        One listing for every role. The visible set is decided by row-level security
        from the verified token, not by the path: an agent sees their own calls, a
        supervisor their team's, qa/compliance/admin the whole tenant, and a bank
        client only flagged calls.

        This exists because the role matrix grants qa, compliance and admin "all
        tenant calls" and the `me` and `teams` listings cannot express that — the
        call explorer is the workhorse screen and a QA reviewer has to be able to
        reach any call in the tenant.
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: user_uid, in: query, schema: { type: string }, description: narrows within the caller's scope; it can never widen it }
        - { name: team_id, in: query, schema: { type: string, format: uuid } }
        - { name: disposition, in: query, schema: { $ref: '#/components/schemas/Disposition' } }
        - { name: has_flags, in: query, schema: { type: boolean } }
        - { name: q, in: query, schema: { type: string }, description: full-text over transcripts in scope }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: page of calls
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CallPage' }
  /v1/calls/{id}:
    get:
      operationId: getCall
      tags: [calls]
      summary: One call, if it is visible at the caller's scope
      description: |
        Returns 404 rather than 403 for a call outside the caller's scope: whether a
        given call id exists in another team is itself information. Every read is
        written to the audit log — a compliance product has to be able to answer who
        listened to a borrower's call.
      parameters: [{ $ref: '#/components/parameters/CallId' }]
      responses:
        '200':
          description: call detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CallDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Identity Platform ID token, verified server-side against Google's public keys.
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    CallId:
      name: id
      in: path
      required: true
      schema: { type: string, description: ULID or UUID rendering of the call id }
  responses:
...
    NotFound:
      description: not found, or not visible at the caller's scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
...
    Severity:
      type: string
      enum: [low, medium, high, critical]
    FlagStatus:
      type: string
      enum: [open, assigned, upheld, dismissed]
    Disposition:
      type: string
      enum: [ptp, refusal, dispute, wrong_number, no_contact, callback_requested,
             partial_payment, escalation, other]
    CaptureTier:
      type: string
      enum: [A, B]
...
    CallSummary:
      type: object
      required: [id, started_at, user_uid, capture_tier, status, flag_count]
      properties:
        id: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, nullable: true }
        user_uid: { type: string }
        account_ref: { type: string, nullable: true }
        direction: { type: string, nullable: true }
        capture_tier: { $ref: '#/components/schemas/CaptureTier' }
        status:
          type: string
          enum: [ingesting, transcribing, analyzing, complete, failed, discarded]
        disposition: { $ref: '#/components/schemas/Disposition' }
        summary: { type: string, nullable: true }
        ptp: { $ref: '#/components/schemas/Ptp' }
        sentiment_delta: { type: number, nullable: true }
        flag_count: { type: integer }
        max_severity: { $ref: '#/components/schemas/Severity' }
    CallPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CallSummary' }
        next_cursor: { type: string, nullable: true }
    CallDetail:
      allOf:
        - $ref: '#/components/schemas/CallSummary'
        - type: object
          properties:
            transcript:
              type: array
              description: speaker-separated turns, merged from the two channels by timestamp
              items: { $ref: '#/components/schemas/TranscriptTurn' }
            sentiment: { $ref: '#/components/schemas/SentimentSeries' }
            flags:
              type: array
              items: { $ref: '#/components/schemas/Flag' }
            talk_ratio: { type: number, nullable: true }
            interruptions: { type: integer, nullable: true }
            next_action: { type: string, nullable: true }
            audio_url:
              type: string
              nullable: true
              description: |
                Short-lived signed URL. Null when the caller's role or the tenant's
                `allow_agent_audio_playback` policy forbids playback. Every issuance
                writes an audit entry.
            analysis:
              type: object
              properties:
                model: { type: string }
                prompt_version: { type: string }
                asr_provider: { type: string }
                asr_version: { type: string }
    TranscriptTurn:
      type: object
      required: [channel, speaker, start_ms, end_ms, text]
      properties:
        channel: { type: integer, enum: [0, 1] }
        speaker: { type: string, enum: [borrower, agent] }
        start_ms: { type: integer }
        end_ms: { type: integer }
        text: { type: string }
        confidence: { type: number, nullable: true }
    SentimentSeries:
      type: object
      properties:
        far:
          type: array
          items: { $ref: '#/components/schemas/SentimentPoint' }
        near:
          type: array
          items: { $ref: '#/components/schemas/SentimentPoint' }
        far_open: { type: number }
        far_close: { type: number }
        delta: { type: number }
    SentimentPoint:
      type: object
      required: [t_ms, v]
      properties:
        t_ms: { type: integer }
        v: { type: number, minimum: -1, maximum: 1 }
    Ptp:
      type: object
      nullable: true
      description: |
        Null when no promise was extracted. When present at all, `present` is
        required, so "no PTP" has exactly one representation rather than three.
      required: [present]
      properties:
        present: { type: boolean }
        amount_paise: { type: integer, format: int64, nullable: true }
        due_date: { type: string, format: date, nullable: true }
        confidence: { type: number }
        evidence_span_ms:
          type: array
          items: { type: integer }
          minItems: 2
          maxItems: 2
        agent_confirmed: { type: boolean, nullable: true }
        agent_amount_paise: { type: integer, format: int64, nullable: true }
        agent_due_date: { type: string, format: date, nullable: true }
...
    Flag:
      type: object
      required: [id, call_id, rule_id, severity, tier, status]
      properties:
        id: { type: string, format: uuid }
        call_id: { type: string }
        rule_id: { type: string }
        rule_set_version: { type: integer }
        severity: { $ref: '#/components/schemas/Severity' }
        tier: { type: integer, enum: [1, 2], description: 1 deterministic, 2 LLM judge }
        span_start_ms: { type: integer, nullable: true }
        span_end_ms: { type: integer, nullable: true }
        evidence_text: { type: string, nullable: true }
        judge_rationale: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/FlagStatus' }
        reviewer_uid: { type: string, nullable: true }
        agent_response: { type: string, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
...